rename image evalautors script to snakecase
