+++ mteb/evaluation/evaluators/image/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-from .Any2AnyMultiChoiceEvaluator import Any2AnyMultiChoiceEvaluator
-from .Any2AnyRetrievalEvaluator import Any2AnyRetrievalEvaluator
-from .ImageTextPairClassificationEvaluator import ImageTextPairClassificationEvaluator
+from .any2any_multichoice_evaluator import Any2AnyMultiChoiceEvaluator
+from .any2any_retrieval_evaluator import Any2AnyRetrievalEvaluator
+from .imagetext_pairclassification_evaluator import ImageTextPairClassificationEvaluator
 
 __all__ = [
-    "Any2AnyMultiChoiceEvaluator",
-    "Any2AnyRetrievalEvaluator",
-    "ImageTextPairClassificationEvaluator",
+    "any2any_multichoice_evaluator",
+    "any2any_retrieval_evaluator",
+    "imagetext_pairclassification_evaluator",
 ]
