+++ mteb/evaluation/evaluators/image/any2any_multichoice_evaluator.py
