+++ mteb/evaluation/evaluators/image/any2any_retrieval_evaluator.py
