+++ mteb/evaluation/evaluators/image/imagetext_pairclassification_evaluator.py
